C/Reverse Polish Notation Calculator
Since it does not need to use dynamic memory, this is actually less complex than JSON parsing,
and i was able to complete in just some hours (granted, I did not implement number parsing,
but it was not really the most time-consuming part of JSON parsing anyway). If someone were
to also do this challenge, you might want to implement this one before the JSON parser.